+++ remote-runner/blocks.go
@@ -5,7 +5,6 @@ import (
 	"os"
 	"remote-runner/config"
 	"strings"
-	"time"
 )
 
 const BlockBufferSizeInBytes = 1024 * 1024 * 4 // 4MB
@@ -26,10 +25,29 @@ const (
 )
 
 // ** Block ** //
+func ClearAllBlocksInDir() error {
+	blockLocationRoot := config.GetConfigs().GetConfigsWithDefault("harpy.remoteRunner.blockMountLocation", "./_blocks")
+	files, err := os.ReadDir(blockLocationRoot)
+	if err != nil {
+		return err
+	}
+	for _, file := range files {
+		err := os.Remove(blockLocationRoot + "/" + file.Name())
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
 type Block struct {
-	BlockID        string
-	BlockLocation  string
-	LastAccessTime int64
+	BlockID       string
+	BlockLocation string
+}
+
+func (b *Block) CheckBlockExists() bool {
+	_, err := os.Stat(b.BlockLocation)
+	return err == nil
 }
 
 func (b *Block) Cleanup() error {
@@ -46,9 +64,11 @@ func (b *Block) Cleanup() error {
 	} else if b.BlockLocation == "../" {
 		return errors.New("block location is parent directory")
 	}
-	err := os.Remove(b.BlockLocation)
-	if err != nil {
-		return err
+	if b.CheckBlockExists() {
+		err := os.Remove(b.BlockLocation)
+		if err != nil {
+			return err
+		}
 	}
 	return nil
 }
@@ -126,7 +146,6 @@ func (br *BlockReader) StartReading() error {
 		return err
 	}
 	br.BlockEndAt = int(blockEndAt)
-	br.Block.LastAccessTime = time.Now().Unix()
 	return nil
 }
 
@@ -158,7 +177,6 @@ func (bw *BlockWriter) StartWriting() error {
 	}
 	bw.IOHandler = file
 	bw.BlockState = BlockWriterStateStreaming
-	bw.Block.LastAccessTime = time.Now().Unix()
 	return nil
 }
